+++ src/visualScanner.js
@@ -1,11 +1,11 @@
-import { updateLatestVisualSignature } from './storage.js?rmv=1.4.9-loaderdiag1';
-import { consumeInjectedFeedbackForSuccessfulRabbitMirror } from './feedbackCat.js?rmv=1.4.9-loaderdiag1';
-import { getSettings } from './settings.js?rmv=1.4.9-loaderdiag1';
+import { updateLatestVisualSignature } from './storage.js?rmv=1.4.9-lightboot1';
+import { consumeInjectedFeedbackForSuccessfulRabbitMirror } from './feedbackCat.js?rmv=1.4.9-lightboot1';
+import { getSettings } from './settings.js?rmv=1.4.9-lightboot1';
 import {
     captureRabbitMirrorGenerationSnapshots,
     getRabbitMirrorGenerationSnapshot,
     inspectRabbitMirrorGenerationSource,
-} from './generationGuard.js?rmv=1.4.9-loaderdiag1';
+} from './generationGuard.js?rmv=1.4.9-lightboot1';
 import { detectMissingVisualProgram } from './presentationQuality.js?rmv=1.4.30.22';
 
 const TOTO_RE = new RegExp('<toto\\b[^>]*(?:data-rabbit-mirror|data-rabbit-' + 'h' + 'ole)=[\"\']true[\"\'][^>]*>[\\s\\S]*?<\\/toto>', 'i');